hybridrag/requirements.txt
# Core dependencies for LightRAG + DeepLake Hybrid System
lightrag-hku>=0.1.0
deeplake>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
numpy>=1.24.0